+++ PRA_B4_FOTOKIOSK/models/KioskProduct.cs
@@ -10,7 +10,10 @@ namespace PRA_B4_FOTOKIOSK.models
     {
 
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        
+        public double Price { get; set; }
+
 
     }
 }
+
